% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyzeExperiment}
\alias{analyzeExperiment}
\title{Analyzes the experimental data.}
\usage{
analyzeExperiment(ex_obj, round_diget = 3)
}
\arguments{
\item{ex_obj}{The experimental object generated by \code{\link{dataframeToRCaMP}}.}

\item{round_diget}{How to round the results digets. Default is 3.}
}
\value{
An rcamp object with a filled results table.
}
\description{
\code{analyzeExperiment} takes in an rcamp object, and
returns physiological relivant data. This includes:
}
\details{
* Peaks
* Troughs (mins)
* T50, T85, and T90
* Vmax for upstrokes and downstrokes.

TODO Pass the peakfinder alpha to this function.
}
