% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calcualteT50}
\alias{calcualteT50}
\title{Calculate T50 Upstroke or Downstroke}
\usage{
calcualteT50(listTime, peaks, mids, left = T)
}
\arguments{
\item{listTime}{List of time measuments}

\item{left}{Boolean, whether to take the t50 from the left (decay/downstroke) or the right (peak/upstroke)}
}
\description{
Calulates the BPM per sample
}
\examples{
x<- calculateBPM(dat$R1, dat$Time)
}

