% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{Calculate BPM}
\alias{Calculate BPM}
\alias{calcualteBPM}
\title{Calculate Beats Per Minute}
\usage{
calcualteBPM(listInt, listTime)
}
\arguments{
\item{listInt}{List of GcAMP measurments}

\item{listTime}{List of time measuments}
}
\description{
Calulates the BPM per sample
}
\examples{
x <- calculateBPM(dat$R1, dat$Time)
}

