% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecking.R
\name{checkTime}
\alias{checkTime}
\title{Check that the vector looks like a time vector,
by checking that each value constently increases
from every value to the next.}
\usage{
checkTime(listTime)
}
\description{
Check that the vector looks like a time vector,
by checking that each value constently increases
from every value to the next.
}
