% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_new_rcamp}
\alias{create_new_rcamp}
\title{Create a new instance of an experiment object, which will initially hold the
raw data from the experiment. The 'Results' attribute will be filled in later
as the program runs. The idea is that each object will hold both the raw
data in the case of reanalysis, and the analysis results for easy comparison
between other experiment objects.}
\usage{
create_new_rcamp()
}
\description{
Objects will have 4 attibutes:
names : Names for each column.
time : A numeric vector containing the timing information (a ms for each
measured value)
data : A matrix containing the intenity values for each measured point
results : a list containing all measured data
}
