% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dataframeToRCaMP}
\alias{dataframeToRCaMP}
\title{Converts a cleaned CSV to an rcamp object.}
\usage{
dataframeToRCaMP(raw_data, time_index = 1, smooth_fxn = FALSE,
  smooth_n = 20, timing_alpha = 2, time_units = "miliseconds",
  time_override = F)
}
\arguments{
\item{raw_data}{A dataframe containing the data you wish to turn into an experiment
object. You are welcome to subset the data before passing it in (ie \code{raw_data[,-c(2,5)]}
to exclude the second and third column).}

\item{time_index}{The index of the column containing all of the 'Time' data.}

\item{smooth_fxn}{A Boolean (default \code{False}). Do you want the data, which may or may
not be noisy, to be smoothed using a running average. Recommended.}

\item{smooth_n}{A number to be passed to the running average function. n is the number
of points to average at each window. Default is 20.}

\item{timing_alpha}{A number to pass to the timing checker. This alpha is multiplied to the
median of the differnces of each time measument. If a time measurment exceeds this threshold,
an error is produced. The microscope may have stalled at this point, and furhter investigation is
needed to check the data. Default is 2.}
}
\value{
An rcamp object.
}
\description{
The rcamp object (an S3 object) will hold all the raw data as well as results. This
function checks to make sure the following are true:
}
\details{
* No NA's are contained in the data frame.

* The time_index is assigned correctly, and no timeing errors are detected. Timing
errors are defined as gaps in recording that exceed some threshold (defined as the
median time multiplied by some alpha).

* There are no extranious intensity values.

The resulting object can be passed to the other funcitons in this program.
}
