% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{findMids}
\alias{findMids}
\title{A funciton to identify midpoints (T50) between peaks and troughs. Only the mid points between .
Argunments :}
\usage{
findMids(listInt, peaks = findPeaks(listInt), mins = findMins(listInt),
  Downstroke = T, midPoint = 0.5)
}
\arguments{
\item{peaks}{: Index values coorisponding to peaks. Will run findPeaks() if nothing given.}

\item{mins}{: Intex values coorisponding to troughs. Will run findMins() if nothing given.}

\item{list}{: A list of GcAMP intensity values coorisponding to an experiemntal run.}
}
\value{
Returns a vector of indexes coorisponding to T50 values. It will only look for T50
  values between peaks and troughs, so nothing should be showing up before the index of
  the first peak or after the index of the last peak.
}
\description{
A funciton to identify midpoints (T50) between peaks and troughs. Only the mid points between .
Argunments :
}
