% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{findPeaks}
\alias{findPeaks}
\title{A funciton to identify peaks in beating GcAMP cells.}
\usage{
findPeaks(list, alpha = 0.01, returnValue = F)
}
\arguments{
\item{list}{: A list of GcAMP intensity values coorisponding to an experiemntal run.}

\item{alpha}{: The % of the data to scan in both directions to find peaks. Defaults to 1%.}

\item{returnValue}{: Boolean whether or not to return indexes or intensity values.}
}
\value{
A vector of either indexes or intensity values coorisponding to peaks.
}
\description{
Peaks are called by running through the intensity values and
dipping below the median of all the values. If the trajectory is positive (moving up),
then the program will continually look for the index until the slop does negative. It will
then look for the max value around the place the slope becomes negative.
}
