% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{indexesToValues}
\alias{indexesToValues}
\title{Converts the indexes (useful for storage and ploting) to their coorisponging
values within the list.}
\usage{
indexesToValues(list, indexes)
}
\description{
Converts the indexes (useful for storage and ploting) to their coorisponging
values within the list.
}
