% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFinder.R
\name{plotRawData}
\alias{plotRawData}
\title{A generic funciton to plot intensities over time The data should be cleaned, either
maually or with \code{\link{readZiessData}} so that only the time and the data you
care about are left. \code{plotRawData} will iterate through the rows, printing}
\usage{
plotRawData(raw_data, time_index = 1)
}
\arguments{
\item{raw_data}{A dataframe containing the data you wish to plot.}

\item{time_index}{The index of the column containing all of the 'Time' data.}
}
\value{
None
}
\description{
A generic funciton to plot intensities over time The data should be cleaned, either
maually or with \code{\link{readZiessData}} so that only the time and the data you
care about are left. \code{plotRawData} will iterate through the rows, printing
}
