% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFinder.R
\name{readZiessData}
\alias{readZiessData}
\title{A funciton to handle loading in data from a Zies reading.
Designed to remove the empty "Marker" column, any empty rows after
importing into R, and removing rows that do not contain mean
intensity values}
\usage{
readZiessData(path_to_csv, time_index = 1, grep_keyword = "IntensityMean")
}
\arguments{
\item{path_to_csv}{A csv file containing data gathered from a Ziess experiment.}

\item{time_index}{The index of the column containing all of the 'Time' data.}

\item{grep_keyword}{The keyword passed to grep to fetch the colums for your
particular experiment. By default, \code{grep_keyword} is set to 'IntensityMean', fetching
columns containing mean intensity values.}
}
\value{
A formated dataframe.
}
\description{
A funciton to handle loading in data from a Zies reading.
Designed to remove the empty "Marker" column, any empty rows after
importing into R, and removing rows that do not contain mean
intensity values
}
