% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{subset_intensity_data}
\alias{subset_intensity_data}
\title{greps the mean intensity values, or whatever othe values you want.}
\usage{
subset_intensity_data(exp_dataset, grep_keyword = "IntensityMean")
}
\description{
greps the mean intensity values, or whatever othe values you want.
}
