% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trimData}
\alias{trimData}
\title{Trim seconds off the begining or end of a run}
\usage{
trimData(raw_data, time_index, trim_start = raw_data[, time_index][1],
  trim_end = raw_data[, time_index][length(raw_data)])
}
\arguments{
\item{raw_data}{A dataframe containing the data you wish to plot.}

\item{time_index}{The index of the column containing all of the 'Time' data.}

\item{trim_start}{Time value to trim to, starting from the left. Default is no trimming.}

\item{trim_end}{Time value to trim to, starting from the right. Default is no trimming.}
}
\description{
Removes data from either end of a run. The value passed to \code{trim_start} will remove values
from the left until the time matches \code{trim_start}. The value passed to \code{trim_end} will remove values
from the right until the time matches \code{trim_end}. For instance, if my time was the vector
\code{c(1,2,3,4,5,6,7)} and I passed \code{trim_start = 2} \code{trim_end = 5}, the output would be
\code{c(2,3,4,5)}. Passing only \code{trim_start = 3} would output \code{c(3,4,5,6,7)}
}
