% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFinder.R
\name{runTestGraph}
\alias{runTestGraph}
\title{Graphs peak calls, troughs (mins), and T50, T85, and T90 reference points,
color coaded and coded based on upstroke or downstroke values.}
\usage{
runTestGraph(exp_obj, name_of_column = exp_obj$names)
}
\arguments{
\item{exp_obj}{The experimental object generated by
\code{\link{dataframeToExperiment}} and analyzed in \code{\link{analyzeExperiment}}.}

\item{name_of_column}{A list or a string that matches the column names you wish to
graph. Defalts to all columns.}
}
\value{
None
}
\description{
Graphs peak calls, troughs (mins), and T50, T85, and T90 reference points,
color coaded and coded based on upstroke or downstroke values.
}
